==> analysis/02_Descriptives.Rmd <==
---
title: "Tech 1. Data Prep & Descriptives"
author: "Nate Breznau"
date: "3/4/2021"
output: 
  html_document:
    keep_md: true

---

```{r setup, include=FALSE}

knitr::opts_chunk$set(warning = F, 
                      echo = T
                      # dev = "ragg_png", 
                      # dev.args = list(png =list(res = 144, width = 3000, height = 2100)),
                      # fig.path = here::here("results\fig-"))
                      )

# need to do this if not already installed


# devtools::install_github("sjmgarnier/viridisLite")


pacman::p_load("tidyverse","kableExtra","rworldmap","ragg", "viridisLite")



```

Data are loaded from the `/src` folder.

**TO DO** Make a list of all sources here (or somewhere)

### Data

Note that we have the raw WHO data (suicide_total_m, suicide_total_f), and
we have the Max Roser 'Our World in Data' data which has WHO plus several supplementary sources.

Note that we use "df_raw" here but we are using "df_imp" the imputed data.

Clean data

```{r load}
df_raw <- readRDS(here::here("data/df_imp_raw.Rds"))

# some cleaning, WHO variables, plus total suicide rate from OWD
df_raw <- df_raw %>%
  mutate(suicide_total_owd = suicide_rate_total_owd*10,
         suicide_rate_female_owd = suicide_rate_female_owd*10,
         suicide_rate_male_owd = suicide_rate_male_owd*10) %>%
  filter(suicide_rate_male_owd != "NaN")
```

#### Descriptives

```{r desc}
df_desc <- df_raw %>%
  filter(year > 2009) %>%
  group_by(iso3c) %>%
  summarise(suicide_pcm_owd = mean(suicide_total_owd, na.rm = T),
            suicide_m_pcm_owd = mean(suicide_rate_male_owd, na.rm = T),
            suicide_f_pcm_owd = mean(suicide_rate_female_owd, na.rm = T))

# calculate change score
df_desc90 <- df_raw %>%
  filter(year < 1995) %>%
  group_by(iso3c) %>%
  summarise(suicide_pcm_owd90 = mean(suicide_total_owd, na.rm = T),
            suicide_m_pcm_owd90 = mean(suicide_rate_male_owd, na.rm = T),
            suicide_f_pcm_owd90 = mean(suicide_rate_female_owd, na.rm = T)) %>%
  select(iso3c, suicide_pcm_owd90, suicide_m_pcm_owd90, suicide_f_pcm_owd90)

df_desc <- df_desc %>%
  left_join(df_desc90, by = c("iso3c")) %>%
  mutate(change_total = round(((suicide_pcm_owd90/suicide_pcm_owd)*100)-100),0)

kable_styling(kable(df_desc, digits = 1))
```

### World Maps

#### Data Prep

```{r cutdata}
# Trim into 7 equal sized groups

df_desc <- df_desc %>%
  mutate(suicide_pcm_owd_c = Hmisc::cut2(suicide_pcm_owd, g=7),
         suicide_m_pcm_owd_c = Hmisc::cut2(suicide_m_pcm_owd , g=7),
         suicide_f_pcm_owd_c = Hmisc::cut2(suicide_f_pcm_owd , g=7),
         change_total_c = Hmisc::cut2(change_total, g=7)) %>%
  group_by(suicide_pcm_owd_c) %>% #for labelling
  mutate(total_min = min(suicide_pcm_owd),
         total_max = max(suicide_pcm_owd)) %>%
  ungroup()

# males label
df_desc <- df_desc %>%
  group_by(suicide_m_pcm_owd_c) %>%
  mutate(male_min = min(suicide_m_pcm_owd),
         male_max = max(suicide_m_pcm_owd)) %>%
  ungroup()

# females label
df_desc <- df_desc %>%
  group_by(suicide_f_pcm_owd_c) %>%
  mutate(female_min = min(suicide_f_pcm_owd),
         female_max = max(suicide_f_pcm_owd)) %>%
  ungroup()

# change label
df_desc <- df_desc %>%
  group_by(change_total_c) %>%
  mutate(change_min = min(change_total),
         change_max = max(change_total)) %>%
  ungroup()

# label file
df_label <- matrix(nrow = 8, ncol = 7)
df_label[1,] <- sort(round(unique(df_desc$total_min), 0))
df_label[2,] <- sort(round(unique(df_desc$total_max), 0))
df_label[3,] <- sort(round(unique(df_desc$male_min), 0))
df_label[4,] <- sort(round(unique(df_desc$male_max), 0))
df_label[5,] <- sort(round(unique(df_desc$female_min), 0))
df_label[6,] <- sort(round(unique(df_desc$female_max), 0))
df_label[7,] <- sort(round(unique(df_desc$change_min), 0))
df_label[8,] <- sort(round(unique(df_desc$change_max), 0)) 

total_label <- c(paste0("<",df_label[2,1]), paste0(df_label[1,2],"-",df_label[2,2]), paste0(df_label[1,3],"-",df_label[2,3]), paste0(df_label[1,4],"-",df_label[2,4]), paste0(df_label[1,5],"-",df_label[2,5]), paste0(df_label[1,6],"-",df_label[2,6]),
                 paste0(">",df_label[1,7]))

male_label <- c(paste0("<",df_label[4,1]), paste0(df_label[3,2],"-",df_label[4,2]), paste0(df_label[3,3],"-",df_label[4,3]), paste0(df_label[3,4],"-",df_label[4,4]), paste0(df_label[3,5],"-",df_label[4,5]), paste0(df_label[3,6],"-",df_label[4,6]),
                 paste0(">",df_label[3,7]))

female_label <- c(paste0("<",df_label[6,1]), paste0(df_label[5,2],"-",df_label[6,2]), paste0(df_label[5,3],"-",df_label[6,3]), paste0(df_label[5,4],"-",df_label[6,4]), paste0(df_label[5,5],"-",df_label[6,5]), paste0(df_label[5,6],"-",df_label[6,6]),
                 paste0(">",df_label[5,7]))

change_label <- c(paste0("<",df_label[8,1],"%"), paste0(df_label[7,2],"- ",df_label[8,2]), paste0(df_label[7,3],"-",df_label[8,3]), paste0(df_label[7,4],"-",df_label[8,4]), paste0(df_label[7,5],"-",df_label[8,5]), paste0(df_label[7,6],"-",df_label[8,6]),
                 paste0(">",(df_label[8,6])+1,"%"))
                 
df_desc <- df_desc %>%
  mutate(suicide_total_factor = factor(suicide_pcm_owd_c, labels = total_label),
         suicide_male_factor = factor(suicide_m_pcm_owd_c, labels = male_label),
         suicide_female_factor = factor(suicide_f_pcm_owd_c, labels = female_label),
         suicide_change_factor = factor(change_total_c, label = change_label))

# setup map
map_desc <- joinCountryData2Map(df_desc,
                                joinCode = "ISO3",
                                nameJoinColumn = "iso3c")

map_desc <- subset(map_desc, continent != "Antarctica")
```

#### Total Suicides

```{r Fig1_total_map}


agg_png(filename = here::here("results/Fig1_total_map.png"), res = 144, width = 3000, height = 2100)
map01 <- mapCountryData(map_desc, 
                        nameColumnToPlot = "suicide_total_factor", 
                        addLegend = T,
                        mapTitle = "Suicides Per Million Capita",
                        missingCountryCol = "ivory",
                        #colourPalette = c('olivedrab4','olivedrab3','olivedrab1', 'goldenrod1','goldenrod3','brown2','brown4'),
                        colourPalette = viridisLite::mako(n = 7, direction = -1),
                        catMethod = 'categorical')
dev.off()

knitr::include_graphics(here::here("results/Fig1_total_map.png"))
```


#### Males

```{r Fig1_males_map}

agg_png(filename = here::here("results/Fig1_males_map.png"), res = 144, width = 3000, height = 2100)
map01 <- mapCountryData(map_desc, 
                        nameColumnToPlot = "suicide_male_factor", 
                        addLegend = T,
                        mapTitle = "Male Suicides Per Million Male Capita",
                        missingCountryCol = "ivory",
                        #colourPalette = c('olivedrab4','olivedrab3','olivedrab1', 'goldenrod1','goldenrod3','brown2','brown4'),
                        colourPalette = viridisLite::mako(n = 7, direction = -1),
                        catMethod = 'fixedWidth')
dev.off()

knitr::include_graphics(here::here("results/Fig1_males_map.png"))
```


#### Females

```{r Fig1_females_map}

agg_png(filename = here::here("results/Fig1_females_map.png"), res = 144, width = 3000, height = 2100)
map01 <- mapCountryData(map_desc, 
                        nameColumnToPlot = "suicide_female_factor", 
                        addLegend = T,
                        mapTitle = "Female Suicides Per Million Female Capita",
                        missingCountryCol = "ivory",
                        #colourPalette = c('olivedrab4','olivedrab3','olivedrab1', 'goldenrod1','goldenrod3','brown2','brown4'),
                        colourPalette = viridisLite::mako(n = 7, direction = -1),
                        catMethod = 'fixedWidth')
dev.off()

knitr::include_graphics(here::here("results/Fig1_females_map.png"))
```

### Change 1990-2018

```{r Fig2_change}

agg_png(filename = here::here("results/Fig2_change.png"), res = 144, width = 3000, height = 2100)
map01 <- mapCountryData(map_desc, 
                        nameColumnToPlot = "suicide_change_factor", 
                        addLegend = T,
                        mapTitle = "Percent Change in Suicide Rate from 1990-2018",
                        missingCountryCol = "ivory",
                        #colourPalette = c('olivedrab4','olivedrab3','olivedrab1', 'goldenrod1','goldenrod3','brown2','brown4'),
                        colourPalette = viridisLite::mako(n = 7, direction = -1),
                        catMethod = 'fixedWidth')
dev.off()

knitr::include_graphics(here::here("results/Fig2_change.png"))
```


### Descriptive Table

```{r descriptives, include=FALSE}

tbl1 <- df_raw %>%
  ungroup() %>% 
  summarise_at(vars(suicide_rate_total_owd, suicide_rate_male_owd, suicide_rate_female_owd, unemp_i, alc_all_types_i, fertility_i,
                    pop65,crisis_idx_i, atwar_i, gdp_cap_i, regime_i, gini_disp_i, fem_labour_i, christpct_ii, islampct_ii, buddhapct_ii,
                    hindupct_ii, nonreligpct_ii, jewishpct_ii, otherpct_ii), 
               funs(mean(., na.rm = T), sd(., na.rm = T), min(., na.rm = T), max(., na.rm = T), N = sum(!is.na(.), na.rm = T)))

splt <- length(tbl1[1,])/5

tbl1 <- as.data.frame(t(tbl1))

tbl_1 <- as.data.frame(matrix(nrow = splt, ncol = 6))

colnames(tbl_1) <- c("Variable", "Mean", "SD", "Min", "Max", "N")

# get proper names eventually
tbl_1[,1] <- rownames(tbl1)[1:splt]

tbl_1[,2] <- round(tbl1[1:splt,1],2)
tbl_1[,3] <- round(tbl1[(splt+1):(splt*2),1],2)
tbl_1[,4] <- round(tbl1[((splt*2)+1):(splt*3),1],2)
tbl_1[,5] <- round(tbl1[((splt*3)+1):(splt*4),1],2)  
tbl_1[,6] <- round(tbl1[((splt*4)+1):(splt*5),1],0)  

kable_styling(kable(tbl_1))

```

```{r descriptives2}

# more elegant solution

df_raw %>%
  ungroup() %>% 
  summarise_at(vars(suicide_rate_total_owd, suicide_rate_male_owd, suicide_rate_female_owd, unemp_i, alc_all_types_i, fertility_i,
                    pop65,crisis_idx_i, atwar_i, gdp_cap_i, regime_i, gini_disp_i, fem_labour_i, christpct_ii, islampct_ii, buddhapct_ii,
                    hindupct_ii, nonreligpct_ii, jewishpct_ii, otherpct_ii), 
               funs(mean(., na.rm = T), sd(., na.rm = T), min(., na.rm = T), max(., na.rm = T), N = sum(!is.na(.), na.rm = T))) %>% 
  # So far same as above
  # Create stats columns automatically with ".value" and regex pattern
  pivot_longer(cols = everything(), names_to = c("Variable", ".value"), 
               # First pattern = Variable name, second = statistic
               names_pattern = "(.+)_(mean|sd|min|max|N)") %>% 
  # Capitalize first letter
  rename_with(~str_to_title(.)) %>% 
  kable(digits = 2) %>% 
  kable_styling()

```
